Adding fuse-libs, pth-devel and libev to the toplevel list
fuse-libs is a dependency of e2fsprogs; including it prevents the
e2fsprogs split entirely, plus we might be shipping FS utilities anyway.

pth-devel is a dependency of libassuan-devel, fixing a broken libassuan
API.

libev provides a backend for libverto and prevents the nghttp2 package
split; nghttp2 is a runtime dependency of httpd and this split would
bite us soon.

None of those, nor any of their subpackages, require anything we're not
already shipping.

Signed-off-by: Petr Šabata <contyk@redhat.com>